Added conductivity ph contrib, added second direction
In the particle-hole vertex contribution I used a trick
to avoid a double k-point sum

The left hand side (Greens functions) is sampled on the q-grid
while the right hand side is sampled along with a shift of the outer iq loop
this essentially creates the same double k-sum
where the load is distributed more evenly on all the prosses.

Second direction: instead of enforcing xx yy and zz
we calculate all possible contributions.
This is necessary in non-orthogonal lattice systems.